Use relative E distances
modify gcode and settings for relative e distances
prepare for Alpha 13 preview
adjust extruder clearances
from Slic3r manual https://manual.slic3r.org/advanced/sequential-printing
Extruder Clearance
Care has to be taken that the nozzle and extruder does not interfere with already printed parts. Slic3r should warn if it detects the nozzle or extruder will collide with a part, but double check that the layout of the parts will not cause problems. The Extruder clearance parameters help Slic3r detect potential collisions:

Radius - The clearance that should be given around the extruder. Take care if the extruder is not mounted centrally - take the largest safe value.

Height - The vertical distance between the nozzle tip and the X axis rods, or lowest part which may interfere with a finished print.