Use Aligned for stock z-seam
Aligned will align all of the z-seams, this is a better stock setting.
Nearest, does the same thing but on a layer-by-layer basis which on complex models and cyliderical objects results in zits and blobs all over the print.

additional work is needed to refine the z-seam settings with coast and retract to reduce the seam.