bed model is now useable
Modified bed model is now usable as a representative of the actual bed model. The SVG texture model is still problamatic and not behaving as expected. SVG issue may be related to saving/exporting from adobe illustrator.

this commit also fixes an oversight in the creation of the ditto printing printer options. the bed model is not activated for the ditto printing as PrusaSlicer is putting the print area in the middle of the bed model and not at the origin assigned to the print area.