Add max_volumetric_speed for 0.4mm nozzle
from the BIBO website it was discovvered that the max volumetric speed for the stock MK8 0.4mm nozzle is 24cc/hour. Assuming this is cm^3/hr converting to mm^3/sec we get 6.67mm^3/s for the max volumetric speed.

filament_max_volumetric_speed values are unknown and the values are very high and are coppied from the Creality and Prusa configurations.